+++ src/main/java/au/com/crypto/bot/application/trade/Trader.java
@@ -9,6 +9,6 @@ public interface Trader {
     public void raiseSignal(ApplicationControllers ac, long existingSignalId, double price, String symbol,
                             String tradeType,
                             String positionType, Strategies.ConditionsGroup conditionsGroup, String strategyPairName,
-                            Map<String, String> props, String market, int contracts, long exchangeId, long marketEventId,
+                            Map<String, String> props, String market, double contracts, long exchangeId, long marketEventId,
                             String exchangeName);
 }
